Move BracketEnvironment and TestEncryptAMIBackwardsCompatibility
Address some refactoring issues that came up as a result of work on
BracketEnvironment.  Commit these changes separately, so that they don't
muddy up the followup commmit.

Move the BracketEnvironment from the encryptor_service package into the
brkt_cli package.  It's a higher-level system configuration construct
and doesn't apply to the encryptor instance.

Move add_brkt_env_to_brkt_config() from util to brkt_cli, also because
it's a higher-level concept.  util should be for lower-level code.

Move TestEncryptAMIBackwardsCompatibility into the AWS codebase, since
it is AWS-specific.
