+++ brkt_cli/gce/update_gce_image.py
@@ -16,7 +16,8 @@ import logging
 
 from brkt_cli.gce import encrypt_gce_image
 from brkt_cli.gce.gce_service import gce_metadata_from_userdata
-from brkt_cli.util import add_brkt_env_to_brkt_config, Deadline
+from brkt_cli.util import Deadline
+from brkt_cli import add_brkt_env_to_brkt_config
 
 from brkt_cli.encryptor_service import wait_for_encryption
 from brkt_cli.encryptor_service import wait_for_encryptor_up
