+++ test.py
@@ -12,7 +12,6 @@
 # License for the specific language governing permissions and
 # limitations under the License.
 import importlib
-import inspect
 import json
 import tempfile
 import time
@@ -21,63 +20,10 @@ import unittest
 import yaml
 
 import brkt_cli
-import brkt_cli.aws
 import brkt_cli.util
 from brkt_cli import proxy
-from brkt_cli.aws import (
-    encrypt_ami
-)
 from brkt_cli.proxy import Proxy
 from brkt_cli.validation import ValidationError
-
-
-class TestEncryptAMIBackwardsCompatibility(unittest.TestCase):
-
-    def test_attributes(self):
-        required_attributes = (
-            'AMI_NAME_MAX_LENGTH',
-            'DESCRIPTION_SNAPSHOT',
-            'NAME_ENCRYPTOR',
-            'NAME_METAVISOR_ROOT_VOLUME',
-            'NAME_METAVISOR_GRUB_VOLUME',
-            'NAME_METAVISOR_LOG_VOLUME'
-        )
-        for attr in required_attributes:
-            self.assertTrue(
-                hasattr(encrypt_ami, attr),
-                'Did not find attribute encrypt_ami.%s' % attr
-            )
-
-    def test_method_signatures(self):
-        required_method_signatures = (
-            ('append_suffix',
-             ['name', 'suffix', 'max_length']),
-            ('clean_up',
-             ['aws_svc', 'instance_ids', 'security_group_ids']),
-            ('get_encrypted_suffix', []),
-            ('snapshot_encrypted_instance',
-             ['aws_svc', 'enc_svc_cls', 'encryptor_instance',
-              'encryptor_image', 'legacy']),
-            ('register_ami',
-             ['aws_svc', 'encryptor_instance', 'encryptor_image', 'name',
-              'description', 'mv_bdm', 'legacy', 'mv_root_id']),
-            ('wait_for_instance',
-             ['aws_svc', 'instance_id']),
-            ('create_encryptor_security_group', ['aws_svc'])
-        )
-        for mthd, args in required_method_signatures:
-            self.assertTrue(
-                hasattr(encrypt_ami, mthd),
-                'Did not find method encrypt_ami.%s' % mthd
-            )
-            method_ref = encrypt_ami.__dict__[mthd]
-            method_args = inspect.getargspec(method_ref)[0]
-            for arg in args:
-                self.assertIn(
-                    arg, method_args,
-                    'Did not find argument "%s" for method encrypt_ami.%s' % (
-                        arg, mthd)
-                )
 
 
 class TestVersionCheck(unittest.TestCase):
@@ -326,3 +272,19 @@ class TestJWT(unittest.TestCase):
                 base64_malformed_payload, base64_payload, base64_signature)
             with self.assertRaises(ValidationError):
                 brkt_cli.validate_jwt(jwt)
+
+
+class TestBrktEnv(unittest.TestCase):
+
+    def test_add_brkt_env_to_user_data(self):
+        userdata = {}
+        api_host_port = 'api.example.com:777'
+        hsmproxy_host_port = 'hsmproxy.example.com:888'
+        expected_userdata = {
+            'api_host': api_host_port,
+            'hsmproxy_host': hsmproxy_host_port
+        }
+        brkt_env = brkt_cli.parse_brkt_env(
+            api_host_port + ',' + hsmproxy_host_port)
+        brkt_cli.add_brkt_env_to_brkt_config(brkt_env, userdata)
+        self.assertEqual(userdata, expected_userdata)
